+++ components/layout/grid/src/gridItem/index.js
@@ -1,13 +1,15 @@
 import {useMemo} from 'react'
 import PropTypes from 'prop-types'
 import cx from 'classnames'
+import PolymorphicElement from '@s-ui/react-primitive-polymorphic-element'
 
 import {BASE_CLASS, CELL_NUMBERS} from '../settings.js'
 import {getColSpanClassNamesTransform} from './settings.js'
 
 export default function LayoutGridItem({
+  as = 'div',
   children,
-  className = '',
+  className,
   colSpan = 1,
   l,
   lOffset,
@@ -41,12 +43,17 @@ export default function LayoutGridItem({
     className
   )
 
-  return <div className={classNames}>{children}</div>
+  return (
+    <PolymorphicElement as={as} className={classNames}>
+      {children}
+    </PolymorphicElement>
+  )
 }
 
 LayoutGridItem.displayName = 'LayoutGridItem'
 
 LayoutGridItem.propTypes = {
+  as: PropTypes.elementType,
   /**
    * The content of the component.
    */